chore: update license
