+++ dot/network/messages/message.go
@@ -1,3 +1,6 @@
+// Copyright 2024 ChainSafe Systems (ON)
+// SPDX-License-Identifier: LGPL-3.0-only
+
 package messages
 
 // P2PMessage must be implemented by all network messages
