+++ scripts/retrieve_block/retrieve_block_test.go
@@ -6,7 +6,7 @@ package main
 import (
 	"testing"
 
-	"github.com/ChainSafe/gossamer/dot/network"
+	"github.com/ChainSafe/gossamer/dot/network/messages"
 	"github.com/ChainSafe/gossamer/lib/common"
 	"github.com/ChainSafe/gossamer/lib/common/variadic"
 	"github.com/stretchr/testify/require"
@@ -15,41 +15,41 @@ import (
 func TestBuildRequestMessage(t *testing.T) {
 	testcases := []struct {
 		arg      string
-		expected *network.BlockRequestMessage
+		expected *messages.BlockRequestMessage
 	}{
 		{
 			arg: "10",
-			expected: network.NewBlockRequest(
+			expected: messages.NewBlockRequest(
 				*variadic.MustNewUint32OrHash(uint(10)), 1,
-				network.BootstrapRequestData, network.Ascending),
+				messages.BootstrapRequestData, messages.Ascending),
 		},
 		{
 			arg: "0x9b0211aadcef4bb65e69346cfd256ddd2abcb674271326b08f0975dac7c17bc7",
-			expected: network.NewBlockRequest(*variadic.MustNewUint32OrHash(
+			expected: messages.NewBlockRequest(*variadic.MustNewUint32OrHash(
 				common.MustHexToHash("0x9b0211aadcef4bb65e69346cfd256ddd2abcb674271326b08f0975dac7c17bc7")),
-				1, network.BootstrapRequestData, network.Ascending),
+				1, messages.BootstrapRequestData, messages.Ascending),
 		},
 		{
 			arg: "0x9b0211aadcef4bb65e69346cfd256ddd2abcb674271326b08f0975dac7c17bc7,asc,20",
-			expected: network.NewBlockRequest(*variadic.MustNewUint32OrHash(
+			expected: messages.NewBlockRequest(*variadic.MustNewUint32OrHash(
 				common.MustHexToHash("0x9b0211aadcef4bb65e69346cfd256ddd2abcb674271326b08f0975dac7c17bc7")),
-				20, network.BootstrapRequestData, network.Ascending),
+				20, messages.BootstrapRequestData, messages.Ascending),
 		},
 		{
 			arg: "1,asc,20",
-			expected: network.NewBlockRequest(*variadic.MustNewUint32OrHash(uint(1)),
-				20, network.BootstrapRequestData, network.Ascending),
+			expected: messages.NewBlockRequest(*variadic.MustNewUint32OrHash(uint(1)),
+				20, messages.BootstrapRequestData, messages.Ascending),
 		},
 		{
 			arg: "0x9b0211aadcef4bb65e69346cfd256ddd2abcb674271326b08f0975dac7c17bc7,desc,20",
-			expected: network.NewBlockRequest(*variadic.MustNewUint32OrHash(
+			expected: messages.NewBlockRequest(*variadic.MustNewUint32OrHash(
 				common.MustHexToHash("0x9b0211aadcef4bb65e69346cfd256ddd2abcb674271326b08f0975dac7c17bc7")),
-				20, network.BootstrapRequestData, network.Descending),
+				20, messages.BootstrapRequestData, messages.Descending),
 		},
 		{
 			arg: "1,desc,20",
-			expected: network.NewBlockRequest(*variadic.MustNewUint32OrHash(uint(1)),
-				20, network.BootstrapRequestData, network.Descending),
+			expected: messages.NewBlockRequest(*variadic.MustNewUint32OrHash(uint(1)),
+				20, messages.BootstrapRequestData, messages.Descending),
 		},
 	}
