+++ scripts/retrieve_state/retrieve_state.go
@@ -1,0 +1,8 @@
+// Copyright 2024 ChainSafe Systems (ON)
+// SPDX-License-Identifier: LGPL-3.0-only
+
+package retrieve_state
+
+func main() {
+
+}
